+++ DTCD-StyleSystem/src/stories/BaseDateTimePicker.stories.js
@@ -1,7 +1,12 @@
-import BaseDateTimePicker from '../base-components/BaseDateTimePicker/BaseDateTimePicker';
+import BaseDateTimePickerDoc from './docs/BaseDateTimePickerDoc.mdx';
 
 export default {
   title: 'Example/BaseComponents/BaseDateTimePicker',
+  parameters: {
+    docs: {
+      page: BaseDateTimePickerDoc,
+    },
+  },
   argTypes: {
     size: {
       control: {
@@ -12,20 +17,29 @@ export default {
   },
 };
 
-const NAME_COMPONENT = 'date-time-picker';
-
-window.customElements.define(NAME_COMPONENT, BaseDateTimePicker);
-
 const Template = args => {
-  const { visible, value, label, size, disabled, required } = args;
+  const {
+    visible,
+    value,
+    label,
+    size,
+    disabled,
+    required,
+    range,
+    timepicker,
+    timewindows,
+  } = args;
 
-  const picker = document.createElement(NAME_COMPONENT);
+  const picker = document.createElement('date-time-picker');
   picker.visible = visible;
   picker.value = value;
   picker.disabled = disabled;
   picker.required = required;
   picker.label = label;
   picker.size = size;
+  picker.range = range;
+  picker.timepicker = timepicker;
+  picker.timewindows = timewindows;
 
   return picker;
 };
@@ -33,9 +47,12 @@ const Template = args => {
 export const DefaultDateTimePicker = Template.bind({});
 DefaultDateTimePicker.args = {
   visible: false,
-  value: 1653489867960,
+  value: String(Date.now()),
   label: 'Label',
   size: '',
   disabled: false,
   required: false,
+  range: false,
+  timepicker: false,
+  timewindows: false,
 };