+++ DTCD-StyleSystem/src/stories/BaseInput.stories.js
@@ -1,5 +1,3 @@
-import BaseInput from '../base-components/BaseInput/BaseInput';
-
 import BaseInputDoc from './docs/BaseInputDoc.mdx';
 
 export default {
@@ -72,10 +70,6 @@ export default {
   },
 };
 
-const NAME_COMPONENT = 'base-input';
-
-window.customElements.define(NAME_COMPONENT, BaseInput);
-
 const Template = args => {
   const {
     theme = [],
@@ -93,7 +87,7 @@ const Template = args => {
     minlength,
   } = args;
 
-  const input = document.createElement(NAME_COMPONENT);
+  const input = document.createElement('base-input');
 
   if (theme.indexOf('<no modification>') !== -1 || theme === '<no modification>') {
     input.theme = [];