+++ DTCD-StyleSystem/src/stories/DataSourceSelect.stories.js
@@ -1,5 +1,3 @@
-import DataSourceSelect from '../base-components/DataSourceSelect/DataSourceSelect';
-
 import DataSourceSelectDoc from './docs/DataSourceSelectDoc.mdx';
 
 export default {
@@ -11,14 +9,10 @@ export default {
   },
 };
 
-const NAME_COMPONENT = 'datasource-select';
-
-window.customElements.define(NAME_COMPONENT, DataSourceSelect);
-
 const Template = (args) => {
   const {} = args;
 
-  const dataSourceSelect = document.createElement(NAME_COMPONENT);
+  const dataSourceSelect = document.createElement('datasource-select');
 
   return dataSourceSelect;
 };